move writeClient to own file
